CreateEntry implementation
