+++ src/com/rommelrico/filesystem/commands/CreateEntry.scala
@@ -1,6 +1,52 @@
 package com.rommelrico.filesystem.commands
 import com.rommelrico.filesystem.State
+import com.rommelrico.filesystem.files.{DirEntry, Directory}
 
-class CreateEntry(entryName: String) extends Command {
-  override def apply(state: State): State = ???
+class CreateEntry(name: String) extends Command {
+
+  override def apply(state: State): State = {
+    val wd = state.wd
+    if (wd.hasEntry(name)) {
+      state.setMessage(s"Entry $name already exists!")
+    } else if (name.contains(Directory.SEPARATOR)) {
+      state.setMessage(s"$name must not contain separators!")
+    } else if (checkIllegal(name)) {
+      state.setMessage(s"$name: illegal entry name!")
+    } else {
+      doMkdir(name, state)
+    }
+  }
+
+  def checkIllegal(name: String): Boolean = {
+    name.contains(".")
+  }
+
+  def doMkdir(name: String, state: State): State = {
+
+    def updateStructure(currentDirectory: Directory, path: List[String], newEntry: DirEntry): Directory = {
+      if (path.isEmpty) currentDirectory.addEntry(newEntry)
+      else {
+        val oldEntry = currentDirectory.findEntry(path.head).asDirectory
+        currentDirectory.replaceEntry(oldEntry.name, updateStructure(oldEntry, path.tail, newEntry))
+      }
+    }
+
+    val wd = state.wd
+
+    // 1. Getting all the directories in the full Path.
+    val allDirsInPath = wd.getAllFoldersInPath
+
+    // 2. Create a new directory entry in the wd.
+    val newEntry: DirEntry = ???
+
+    // 3. Updating the whole directory structure starting from the root. (IMMUTABLE)
+    val newRoot = updateStructure(state.root, allDirsInPath, newEntry)
+
+    // 4. Find new working directory INSTANCE given wd's full path, in the new directory structure.
+    val newWd = newRoot.findDescendant(allDirsInPath)
+
+    State(newRoot, newWd)
+
+  }
+
 }
